Updated to editor
