--- manifest/manifest.yml
@@ -1,16 +1,0 @@
-version: '1'
-
-# Define all entity types
-entity_types:
-  ProjectFolder:
-    has_designer: false
-    show_properties: false
-  Tunnel:
-    has_designer: true
-    show_properties: true
-
-# On app start, create sample entity
-entities:
-- entity_type: ProjectFolder
-  properties:
-    name: Projects
