+++ requirements.txt
@@ -1,2 +1,2 @@
 viktor==14.0.0
-shapely==1.8.1
+shapely==2.0.1
