+++ viktor.config.toml
@@ -1,1 +1,1 @@
-app_typ="tree"
+app_type="editor"
